bugfix/AOS-6033-deploy-feiler-dersom-man-redigerer-auroraconfig (#165)
